Close on HOME and Clarify Manifest.
The manifest now explicitly states what the Settings Activity parent is
for older API versions. Making it work properly.

The Activity now just closes when it loses focus. Including the home
button.

This is the basic idea for #53 but still needs to have considerations
for opening things like the camera app.