!!! WARNING: this invalidates your current preferences files. Please uninstall application first before launching this branch !!!
Moved the creation of the ConcealCrypto to the Utils class and let the service get it from there. This way the EncryptedPreferences method can get the same ConcealCrypto instance in order to encrypt/decrypt the preference files. Now the preferences are saved in their encrypted state. Next step: encrypting thumbnails.