Add app label morph
Change the way the manifest is decoded.  Right now it just writes it
back exactly the same except for the app label. The xml document still
gets returned to acquiredresource ID's from.

Still doesn't work if the app label refers to a resource. The max name
length is the same as the length of the start name, AndroidStealth, in
bytes right now. It cuts off there.
Shorter than that it gets padded with spaces.