Fix some morphing name issues
Right now the input field for the name has been updated with some input
restrictions.

As well as fixing some issues with not having values for the name and
the icon after morphing.

So far the morphing still fails on some characters that are allowed as
input. Because it messes with file IO. This is: | < > * \
Next step is escaping file IO.